refactor(KERN-1560): move helper modules into helpers/ folder
Organize helper files under sql-to-meerkat/helpers/:
- helpers/aggregate-detection.ts
- helpers/naming.ts
- helpers/ast-utils.ts

helpers.ts remains as barrel re-export — all existing imports unchanged.

work-item: KERN-1560
